QcomModulePkg: Add support for debugging apps
By default the clang toolchain starts the text segment at offset 0x1000
and the offset varies from tool chain to tool chain. In UEFI when the
GenFw tool generates a PECoff image, the tool uses the text segment
offset as '0'. Due to this there would be a symbol mismatch when dll
file is loaded on t32. To avoid this symbol mismatch force the offset to
0x0 using compiler flags.

CRs-Fixed: 980075
Change-Id: I3224f84b2e55c70386d35b24df8d1af645dff102