QcomModulePkg: Add support for Bin B
Add support for Bin B to disable multi media nodes. Move the partial
goods table to the header file defining it in the local block optimizes
out the definition and the memory contents are junk. Also ignore the
nodes which are not present or disabled, as its same as disabling the
nodes

CRs-Fixed: 998065
Change-Id: I83587ed1fcecd5ff8dce947a3d338b2686941254