QcomModulePkg: Clear the LunSet flag & clean up logs
Clear the LunSet flag after the flashing is done because if the next
partition being flashed is not part of the previous lun then flashing
fails with an error saying partition not found. Clearing LunSet flag
lets the fastboot look for the partition in all the luns. Also clean up
the debug logs added as part of fastboot continue command support.

Change-Id: I1b3bc141d2f855946ecaa4c105cb41bfbe340078