QcomModulePkg: Add support for Verified Boot
This change support for verified boot for image authentication and
handle boot states (RED, ORANGE, YELLOW and GREEN) of the device.
Also, send root of trust to the trusted EE based on the image
authentication results.
Initialize the device lock and unlock states and keep the device state
unlocked by default on a non-secure device.

CRs-Fixed: 1042118
Change-Id: I9b3bdbd3d5b95178e3c27d4e85b7b1b5fc08aab4