QcomModulePkg: Update the behavior for A/B update
Set Active slot must clear the active bit and reduce the priority for
the old slot and set the max priority, active bit for the new slot.
After this swap the GUID for all the partitions from old to new.
Also in the swap slot API the local variable updates use non secure
string concatenate applications with wrong size thus overwriting the
variables on the stack, update to use secure version of string apis
with correct size for the copy and concatenation.

Change-Id: I34bc3c90cc60cb6f532d8917306a86aec9fd3e1d