QcomModulePkg: Stop timer when the key is be holding down
The event of keys' detection has higher priority than any other event
in the sysytem, it will cause the enumeration failed when the host PC
detects the device connection if the user is holding down the volume
key. So stop the timer when the key is be holding down.

CRs-Fixed: 1082742
Change-Id: I9a3213ef9e0cbff4f0781f5e548d5d35a4c129d0