QcomModulePkg: Add slot name length check while preparing multi-slot list
Currently, while getting multi-slot list, we do not compare the slot name
length check due to which comparison of slots like 'xxx_xx_a' with 'xxx_b'
and 'xxx_a' holds true and 'xxx' is added two times in the link list. So,
while setting the other slot as active, the swapping of GUID'S is happening
two times for these partitions which means there is no slot switch happening
for these partitions and this can affect booting.

This patch fixes this issue by adding slot name length check in comparison.

Change-Id: Iff859c04ff231b6d804ae3024808afc5cea63e79