QcomModulePkg: Add GUID check along with attribute while updating GPT
In a scenario, where we flash meta on both slots of the partitions the
attribute value of the slots are zero by default. So, when we try to
make a switch to b from a only we mark the active attribute bit set
for b and reset the same for slot a (which are 0 initially ) due to
which while updating partition attributes only the b slots attribute
and GUID's are updated. So both the slots of each partitions have same
GUID. So once we reboot PBL is not able to detect the correct GUID to
boot it from.

This patch fixes this issue by adding more check while updating
attributes.

Change-Id: I5fab0a988c4aa8caf62387ada73c9b0fee120cca