QcomModulePkg: System-As-Root on non-A/B for VB1 targets
According to System-As-Root behaviour, system.img should contain both
system content and ramdisk content, and should be mounted at
root(‘/’). As of today A/B already has system-as-root by default.
Now, we need to add support for the same even on non-A/B.

This change only valid for VB1 targets.

Change-Id: Ic1184aadd9dae975ecf09ab8ea2c08310fc29fb9