QcomModulePkg: Add a recovery DTBO section to boot image header
For non-A/B devices, the recovery partition should be self-sufficient
and not depend on any other partition. Since recovery may need access
to devices added to the DT by overlays, so update the image header to
make sure that the the recovery image also contains the information
from the DTBO image. And fix the compile error after update the image
header.

Change-Id: Ica14bd3cc565b75e7a1cf1c28ae6444652086d77