QcomModulePkg: Clear the old DTB magic in DTBO boot path
Currently, in the booting path if kernel find the DTB magic at
the end of compressed kernel (32-bit), it will overwrite the
passed DeviceTree load address from ABL and it will try to boot
with the found DTB.

Incase of without DTBO, we handle this by clearing out the DTB
magic by setting it to zero. We should also do the same for DTBO
boot path as well.

This patch fix the above mentioned issue.

Change-Id: I27c3c41293c53af67d1794711e75f987c3b9fe2a