QcomModulePkg: Update GPT only when attributes or GUID is changed
In a scenario, when a missing Blkio handle occured and the partition
info in the storage is right, doing attributes and GUID compare will
update the GPT header which will cause the device boot up failed. So
don't update the GPT header if nothing attribute or GUID needed to be
changed.

Change-Id: I093c1b56da0c4315cd312cd031697d7eb10bdd53