edk2: QcomModulePkg: Calculate image offsets based on Header Version
When we disable avb feature, calculate the size of the data needed to be
read from the flash completely. Add suport to read the dtb.img
which is present at the end of the boot image in case of the header
version 2. This change is needed because up until header version 2, dtbs
were appended to the kernel and in kernel size, dtbs were included.

Change-Id: Iad055672e473951fc92c4138c98555e67875aeb0