QcomModulePkg: BootImagev3: Add new header definitions
Unlike the previous boot-image header versions, the version-3
separates out the vendor specific information, such as page size,
kernel load address, device-tree information, and so on, from the
boot-image. In addition to that, version-3 also provides capabilities
to provide its own ramdisk (vendor-ramdisk) and kernel command-line
arguments. These two would be concatenated with the ones in the
boot-image to consider them for booting of the device. Hence,
add the header definitions for boot-image and vendor-boot-image
for boot-image header version 3.

Change-Id: I1923bf04c7fbbe964727f7bca0dc4363e176e1c1