QcomModulePkg: BootImagev3: Update DTB offsets
For boot-image header version three, the DTB is packaged as
a part of vendor-boot image. Hence, update the image buffer,
DTB size and offset accordingly to get the device tree and
parse it.

Also, replace the variable DtbSize with ImageSize in
DTBImgCheckAndAppendDT() as it's a little misleading.
Technically, it's used to hold the size of the Image
so that the DT can be extracted properly from it.

Change-Id: I19cb04ce60f00968baebc7e12adadb8dbad68b04