QcomModulePkg: BootImagev3: Cat vendor ramdisk
For boot-image versions greater than 2, the vendor-boot image
also packs a ramdisk. Before booting the kernel, concatinate
the vendor-ramdisk to the ramdisk at the ramdisk's load address.
This would result in an overlay when the kernel tries to
mount the initrd image. Both .gzip and .cpio formats supports
this kind of overlay through concatination.

Change-Id: Ib2614d9585e9414389af6359983074e8f0a2f912