QcomModulePkg: append force_normal_boot with dp and recovery as boot enabled
For devices where dyanamic partition(DP) is not enabled and system as root
and recovery as boot both flags are enabled, force_normal_boot=1 flag is
not required to boot into normal mode or recovery mode.
For normal bootup system is mounted at / in RO mode hence mkdir and
switch root to /first_stage_ramdisk will fail if force_normal_boot is
passed in commandline.

Change-Id: I79e9ebc18d0b020a3e7280098889a8e1ddf6fc56