QcomModulePkg: Correct the length of node name when adding info to list
Copy the length of current node info to the list rather than the length
of the searching node, wrong info in the list will cause device could
not boot up.

For example: if there are some nodes, such as "regulator-pm8008i-l3",
"regulator-pm8008i-l4". Add only add "regulator-p" to the node list
during searching node's offset. The offset of node in the list will
not be updated correctly when some node's info is updated. Wrong offset
in the list will cause device could not boot up.

Change-Id: I0fb72122c7e88f91c176331a1572734857f60095