QcomModulePkg: Enhance PMIC detection in abl
Currently in ABL maximum supported PMIC is four.
Change to enhance PMIC detection in bootloader to max 0xF.

With this support, pmic will be defined in dt using
qcom,pmic-id-size = <#>;
to represent maximum number of pmic on target.
qcom,pmic-id = <a b c d ...>;
to define pmic ids of pmic on target.

Change-Id: Ia3b6b1efac22b17c7c8c529476ad8bba9ebb940a
Signed-off-by: Mayank Grover <groverm@codeaurora.org>