QcomModulePkg: Skip current pmic-id selection in case of none match
In case pmic-id property is defined for one of the platform
variant and none of the pmic is matches NONE_MATCH bit get set
in DtMatchVal. DtMatchVal with NONE_MATCH is greater
than DtMatchVal for dtb with no pmic-id field defined.
pmic-id with NONE_MATCH should not be picked.
Also, in case pmic revision is none match skip the current
pmic-id.

Change-Id: Ia96bc2ecb7c0b542d6cd8b21710e0f2d96496ed9